Updated services with Jwt

+++ bse/src/main/java/com/demo/bse/serviceImpl/StockServiceImpl.java
@@ -1,10 +1,10 @@
 package com.demo.bse.serviceImpl;
 
+
 import com.demo.bse.entity.Stock;
 import com.demo.bse.entity.StockResponse;
 import com.demo.bse.repository.StockRepository;
 import com.demo.bse.service.StockService;
-import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,19 +17,14 @@ public class StockServiceImpl implements StockService {
     @Autowired
     private StockRepository stockRepository;
 
-//    @Autowired
-//    private NseClient nseClient;
-
-//    @Override
-//    public List<Stock> getAllStocks(){
-//        return stockRepository.findAll();
-//    }
     @Override
     public List<StockResponse> getAllStocks() {
         List<Stock> stocks = stockRepository.findAll();
         // Convert Stock entities to StockResponse objects
         return stocks.stream()
-                .map(stock -> new StockResponse(stock.getName(), stock.getPrice(),stock.getQuantity()))
+                .map(stock -> new StockResponse(stock.getSymbol(),stock.getName(),stock.getPrice()
+                        ,stock.getDayLow(),stock.getDayHigh(),stock.getYearLow(),stock.getYearHigh(),stock.getMarketCap()
+                        ,stock.getExchange(),stock.getOpen(),stock.getPrevClose()))
                 .collect(Collectors.toList());
     }
     @Override
@@ -51,8 +46,13 @@ public class StockServiceImpl implements StockService {
     }
 
     @Override
-    public void deleteStock(long id) {
-        stockRepository.deleteById(id);
+    public boolean deleteStock(long id) {
+        if (stockRepository.existsById(id)) {
+            stockRepository.deleteById(id);
+            return true; // Stock was found and deleted
+        } else {
+            return false; // Stock was not found
+        }
     }
 
     @Override
@@ -64,29 +64,6 @@ public class StockServiceImpl implements StockService {
 
     // Helper method to convert Stock to StockResponse
     private StockResponse convertToStockResponse(Stock stock) {
-        return new StockResponse(stock.getId(), stock.getName(), stock.getPrice(), stock.getQuantity());
+        return new StockResponse();
     }
-
-
-//    @Transactional
-//    public Stock buyStock(Long id, int quantity) {
-//        Stock stock = stockRepository.findById(id)
-//                .orElseThrow(() -> new RuntimeException("Stock not found"));
-//
-//        if (stock.getQuantity() < quantity) {
-//            throw new RuntimeException("Not enough stock available");
-//        }
-//
-//        stock.setQuantity(stock.getQuantity() - quantity);
-//        return stockRepository.save(stock);
-//    }
-//
-//    @Transactional
-//    public Stock sellStock(Long id, int quantity) {
-//        Stock stock = stockRepository.findById(id)
-//                .orElseThrow(() -> new RuntimeException("Stock not found"));
-//
-//        stock.setQuantity(stock.getQuantity() + quantity);
-//        return stockRepository.save(stock);
-//    }
 }
